week_06/pipeline_v2/etl_job/requirements.txt
pymongo
sqlalchemy
psycopg2-binary
pandas
vaderSentiment
requests


